+++ web/src/App.js
@@ -12,6 +12,8 @@ import Preference from './components/Preference';
 import Rewards from './components/Rewards';
 import Tasks from './components/Tasks';
 import Navbar from './components/Navbar';
+import ReadTask from './components/ReadTask';
+
 
 function App() {
   const { isLoggedIn, logout } = useContext(AuthContext);
@@ -56,10 +58,18 @@ function App() {
           }
         />
         <Route
-          path="/tasks"
+          path="/break-task"
           element={
             <ProtectedRoute>
               <Tasks />
+            </ProtectedRoute>
+          }
+        />
+        <Route
+          path="/read-task"
+          element={
+            <ProtectedRoute>
+              <ReadTask />
             </ProtectedRoute>
           }
         />
