+++ web/src/components/Tasks.js
@@ -3,7 +3,8 @@ import TaskBreak from './TaskBreak'; // Import the break task component
 import '../styles/tasks.css';
 import confetti from 'canvas-confetti';
 import axios from 'axios';
-import { Link } from 'react-router-dom'; // Import the confetti library
+import { Link } from 'react-router-dom';
+import SubNavbar from './SubNavbar'; // Import the confetti library
 
 const Tasks = ({ assignedTask }) => {
   const [timer, setTimer] = useState(10); // 10 seconds for demo purposes
@@ -48,23 +49,6 @@ const Tasks = ({ assignedTask }) => {
     }
   }
 
-  function SubNavbar() {
-    return(
-      <div className="task-navbar">
-        <div className="" >
-          <Link className="taskNames" to="/break-task">Break </Link>
-        </div>
-        <div>
-          <Link className="taskNames" to="/read-task">Reading </Link>
-        </div>
-        <div>
-          <Link className="taskNames" to="/exercise-task">Exercise </Link>
-        </div>
-        <div>
-          <Link className="taskNames" to="">Placeholder </Link>
-        </div>
-      </div>)
-  }
   useEffect(() => {
     if (assignedTask) {
       setCurrentTask(assignedTask);
