+++ web/src/styles/ReadTask.css
@@ -1,0 +1,3 @@
+.reading{
+    color: black;
+}
